Solved part two of day12 challenge and increased efficiency by using memoization.

+++ day12-hot-springs/src/helper.rs
@@ -1,12 +1,12 @@
 use std::{fs::File, io::BufRead, io::BufReader};
 
-pub(crate) fn read_input() -> Vec<(Vec<char>, Vec<i32>)> {
+pub(crate) fn read_input() -> Vec<(Vec<char>, Vec<usize>)> {
     let file = File::open("src/input.txt").unwrap();
     let reader = BufReader::new(file);
     reader.lines().map(|l| {
         let l = l.unwrap();
         let line = l.split(" ").collect::<Vec<&str>>();
-        (line[0].chars().collect(), line[1].split(",").map(|s| s.parse::<i32>().unwrap()).collect())
+        (line[0].chars().collect(), line[1].split(",").map(|s| s.parse::<usize>().unwrap()).collect())
     }).collect()
 }
